+++ src/hklpy2/backends/th_tth_q.py
@@ -19,7 +19,7 @@ import math
 from typing import List
 
 from .. import __version__
-from ..misc import IDENTITY_MATRIX_3X3
+from ..utils import IDENTITY_MATRIX_3X3
 from ..exceptions import SolverError
 from ..typing import Matrix3x3
 from ..typing import NamedFloatDict